wrote supercut
it's not as good as i would like, but it works.

fun fact: my local copy of this util directory contained a bash script called supercut for many years. someone else wrote that script, and i copied it to my machine and named it supercut hoping it would do what i wanted. but it did not do what i wanted. so i am writing my own script. and calling that supercut.